Kinda stuck
Trying to figure out how to graph a line in 3D. Found multiple examples of how to make a box, with rotation and perspective, but nothing on how to graph an arc. I figured out how to graph a line, but thats about it. As of now, this is mostly not my code, just seeing what other people have done