We got a plan, SORTA
So the plan is to make an arraylist of a class that draws a single line segment, and from those line segments I will make the arc. However, one problem I realized just now, is that each arc is going to be different, in terms of its length. That means I will have to scale my box to make sure that it fits the arc... Thats gonna make it a lot more complicated.