I think I am done
I managed to do everything I set out to do other than use input. However, I think that is beyond my ability right now to figure out how to have the user type a function or a command and have the program know what it is and what has to be done with it. Overall, it works just as intended